pyproject.toml
[project]
name = "ndbfunctions"
version = "0.1.0"
description = "Minting process for the Neotoma Paleoecology Database's DOIs using the DataCite API and the Neotoma API."
readme = { file = "README.md", content-type = "text/markdown" }
license = "MIT"
license-files = ["LICENSE.md"]
authors = [{ name = "Simon Goring", email = "goring@wisc.edu" }]
maintainers = [{ name = "Simon Goring", email = "goring@wisc.edu" }]
requires-python = ">=3.12"
keywords = ["Postgresql", "metadata", "Open Science"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Hatch",
    "Framework :: Pytest",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: SQL",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
]
dependencies = ["psycopg[binary]>=3.3.2", "python-dotenv>=1.2.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ndbfunctions"]
